move log to utilities folder to be in sync with Als PR
