+++ algopy_testing/src/algopy/__init__.py
@@ -10,12 +10,12 @@ from algopy_testing.models import (
     GTxn,
     ITxn,
     Txn,
-    log,
     urange,
 )
 from algopy_testing.primitives import BigUInt, Bytes, String, UInt64
 from algopy_testing.protocols import BytesBacked
 from algopy_testing.state import GlobalState, LocalState
+from algopy_testing.utilities import log
 
 from . import arc4, gtxn, itxn, op
