+++ algopy_testing/src/algopy_testing/utilities/__init__.py
@@ -1,0 +1,3 @@
+from algopy_testing.utilities.log import log
+
+__all__ = ["log"]
